fix(ngrx): flaky auth karma tests (#122)
Tests for the auth service were flaky because one of the tests was
mutating global state, but tests are not guaranteed to always run
in-order as they are defined (this is determined by the seed). I reset
the token value in auth.service.spec.ts so it's reset before each test
is run to avoid flakiness.